Bug fixing, add context feature (v1.2)
Features:

- Introduce a message passing protocol to enable communication within a paired installation.
- Open Mod can now include the context for an action in the extracts (disabled by default).

Bug fixes.

- Fix an issue that prevented the proper operation of Open Mod.
- Fix an issue where newly observed users were not being added to the content deletion policy enforcement set.

Miscellaneous:

- Update README.md to reflect above changes.

Additional Details:

In order to operate correctly, Open Mod needs to be able to post in the community where you would like your extracts to live.
To achieve this, Open Mod must be additionally installed in that community.
Message passing between the two installations is facilitated by a simple protocol, with a wiki-page serving as the message 'queue'.